feat!: add baidu verification meta tag
This copies _includes/head.html from official chirpy site to this repo,
which will lead to problems on chirpy's upgrade.

The problem, specifically, is that baidu verification meta tag is added
to _includes/head.html, which overrides chirpy's version, so upgrade
won't take effect unless we remove or merge updates to our
_includes/head.html .

This will also be introduced as an issue in my repository.